requirements.txt
# --- Core ---
fastapi==0.116.1
uvicorn==0.35.0
uvloop==0.21.0; sys_platform != "win32"
httptools==0.6.4
starlette==0.47.1
anyio==4.9.0
h11==0.16.0
watchfiles==1.1.0
websockets==15.0.1

# --- Config & Validation ---
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
typing-extensions==4.14.1
annotated-types==0.7.0
python-dotenv==1.1.1

# --- Security & Auth ---
bcrypt==4.3.0
passlib==1.7.4
python-jose==3.5.0
rsa==4.9.1
ecdsa==0.19.1
email-validator==2.2.0
python-multipart==0.0.20

# --- Database ---
sqlalchemy==2.0.41
asyncpg==0.30.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
motor==3.7.1
pymongo==4.15.1
greenlet==3.2.3

# --- Utils ---
requests==2.32.5
httpx==0.28.1
orjson==3.11.0
ujson==5.10.0
lxml==6.0.1
PyYAML==6.0.2
python-slugify==8.0.4
tqdm==4.67.1

# --- Optional (observability/monitoring) ---
sentry-sdk==2.33.0